+++ src/api.ts
@@ -21,6 +21,7 @@ export default class Chromeless<T extends any> implements Promise<T> {
       waitTimeout: 10000,
       remote: false,
       implicitWait: true,
+      launchChrome: true,
 
       ...options,
 
@@ -97,7 +98,8 @@ export default class Chromeless<T extends any> implements Promise<T> {
   }
 
   focus(selector: string): Chromeless<T> {
-    throw new Error('Not implemented yet')
+      this.queue.enqueue({type: 'focus', selector})
+      return this
   }
 
   press(keyCode: number, count?: number, modifiers?: any): Chromeless<T> {
@@ -124,12 +126,14 @@ export default class Chromeless<T extends any> implements Promise<T> {
     throw new Error('Not implemented yet')
   }
 
-  mousedown(): Chromeless<T> {
-    throw new Error('Not implemented yet')
+  mousedown(selector: string): Chromeless<T> {
+      this.queue.enqueue({type: 'mousedown', selector})
+      return this
   }
 
-  mouseup(): Chromeless<T> {
-    throw new Error('Not implemented yet')
+  mouseup(selector: string): Chromeless<T> {
+      this.queue.enqueue({type: 'mouseup', selector})
+      return this
   }
 
   mouseover(): Chromeless<T> {
@@ -138,6 +142,12 @@ export default class Chromeless<T extends any> implements Promise<T> {
 
   scrollTo(x: number, y: number): Chromeless<T> {
     this.queue.enqueue({type: 'scrollTo', x, y})
+
+    return this
+  }
+
+  setHtml(html: string): Chromeless<T> {
+    this.queue.enqueue({type: 'setHtml', html})
 
     return this
   }
@@ -166,6 +176,12 @@ export default class Chromeless<T extends any> implements Promise<T> {
 
   screenshot(): Chromeless<string> {
     this.lastReturnPromise = this.queue.process<string>({type: 'returnScreenshot'})
+
+    return new Chromeless<string>({}, this)
+  }
+
+  getHtml(): Chromeless<string> {
+    this.lastReturnPromise = this.queue.process<string>({type: 'returnHtml'})
 
     return new Chromeless<string>({}, this)
   }